+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Validate credentials to be sent to openvas. [#416](https://github.com/greenbone/ospd-openvas/pull/416)
 - New Credentials for SSH to get su privileges. [#419](https://github.com/greenbone/ospd-openvas/pull/419)
 - Integrate Sentry. [#421](https://github.com/greenbone/ospd-openvas/pull/421)
+- Add dry run support. Generate 10 fake results per host. [#424](https://github.com/greenbone/ospd-openvas/pull/424)
 
 ### Changed
 ### Removed